commit 03

+++ src/main/java/com/portfolio/jepc/Repository/IPersonRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IPersonRepository extends JpaRepository<Person, Long>{
+public interface IPersonRepository extends JpaRepository<Person, Integer>{
    
     public Optional<Person> findByName(String name);
